liberated basemap logic from MapActivity #13
