+++ app/src/main/java/org/redcross/openmapkit/Basemap.java
@@ -1,0 +1,188 @@
+package org.redcross.openmapkit;
+
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
+import android.content.SharedPreferences;
+import android.os.Environment;
+import android.widget.Toast;
+
+import com.mapbox.mapboxsdk.tileprovider.tilesource.MBTilesLayer;
+import com.mapbox.mapboxsdk.tileprovider.tilesource.WebSourceTileLayer;
+import com.mapbox.mapboxsdk.views.MapView;
+
+import java.io.File;
+import java.util.ArrayList;
+
+/**
+ * Created by Clint Cabanero & Nicholas Hallahan on 3/17/15.
+ *
+ * nhallahan@spatialdev.com
+ * ccabanero@spatialdev.com
+ * * * * * * 
+ */
+public class Basemap {
+    
+    private MapActivity mapActivity;
+    private MapView mapView;
+    private Context context;
+    
+    private String selectedMBTilesFile;
+    
+    public Basemap(MapActivity mapActivity) {
+        this.mapActivity = mapActivity;
+        this.mapView = mapActivity.getMapView();
+        this.context = mapActivity.getApplicationContext();
+
+        if (Connectivity.isConnected(context)) {
+            addOnlineDataSources();
+        } else {
+            presentMBTilesOptions();
+        }
+    }
+    
+    private void addOnlineDataSources() {
+        //create OSM tile layer
+        String defaultTilePID = mapActivity.getString(R.string.defaultTileLayerPID);
+        String defaultTileURL = mapActivity.getString(R.string.defaultTileLayerURL);
+        String defaultTileName = mapActivity.getString(R.string.defaultTileLayerName);
+        String defaultTileAttribution = mapActivity.getString(R.string.defaultTileLayerAttribution);
+
+        WebSourceTileLayer ws = new WebSourceTileLayer(defaultTilePID, defaultTileURL);
+        ws.setName(defaultTileName).setAttribution(defaultTileAttribution);
+
+        //add OSM tile layer to map
+        mapView.setTileSource(ws);
+    }
+    
+    public void presentMBTilesOptions() {
+        //shared preferences private to this activity
+        final SharedPreferences sharedPreferences = mapActivity.getPreferences(Context.MODE_PRIVATE);
+        String previousMBTilesChoice = sharedPreferences.getString(mapActivity.getString(R.string.preferenceMBTileChoice), "nopreviouschoice");
+
+        //create an array of all mbtile options
+        final ArrayList<String> mbtilesFileNames = new ArrayList<>();
+
+        //when device is connected, HOT OSM Basemap is the first option
+        if(Connectivity.isConnected(context)) {
+            mbtilesFileNames.add(mapActivity.getString(R.string.hotOSMOptionTitle));
+        }
+
+        //add mbtiles names from external storage
+        File[] mbtiles = ExternalStorage.fetchMBTilesFiles();
+        if (mbtiles != null && mbtiles.length > 0) {
+            for (File file : mbtiles) {
+                String fileName = file.getName();
+                mbtilesFileNames.add(fileName);
+            }
+        }
+
+        //present a dialog of mbtiles options
+        if(mbtilesFileNames.size() > 0) {
+
+            //create dialog of mbtiles choices
+            AlertDialog.Builder builder = new AlertDialog.Builder(mapActivity);
+            builder.setTitle(mapActivity.getString(R.string.mbtilesChooserDialogTitle));
+            CharSequence[] charSeq = (CharSequence[]) mbtilesFileNames.toArray(new CharSequence[mbtilesFileNames.size()]);
+
+            //default mbtiles option is based on previous selections (persisted in shared preferences) or connectivity state of device
+            int defaultRadioButtonIndex = -1; //none radio button selected
+            if(previousMBTilesChoice.equals("nopreviouschoice")) {
+                //if user DID NOT previously choose an mbtiles option...
+                if(Connectivity.isConnected(context)) {
+                    //the first radio button (for HOT OSM) will be selected by default
+                    defaultRadioButtonIndex = 0;
+                    //the default selected option is HOT OSM
+                    selectedMBTilesFile = mbtilesFileNames.get(0); //default choice
+                } else {
+                    defaultRadioButtonIndex = -1; //no selected radio button by default
+                }
+            } else {
+                //if user previously chose an mbtiles option ...
+                for(int i = 0; i < mbtilesFileNames.size(); i++) {
+                    if(mbtilesFileNames.get(i).equals(previousMBTilesChoice)) {
+                        defaultRadioButtonIndex = i;
+                    }
+                }
+            }
+
+            //add choices to dialog
+            builder.setSingleChoiceItems(charSeq, defaultRadioButtonIndex, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int which) {
+                    //user tapped on radio button and changed previous choice or default
+                    String userMBTilesChoice = mbtilesFileNames.get(which);
+
+                    //add user's choice to shared preferences key
+                    SharedPreferences.Editor editor = sharedPreferences.edit();
+                    editor.putString(mapActivity.getString(R.string.preferenceMBTileChoice), userMBTilesChoice);
+                    editor.commit();
+
+                    selectedMBTilesFile = userMBTilesChoice;
+                }
+            });
+
+            //handle OK tap event of dialog
+            builder.setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int id) {
+                    //user clicked OK
+                    if(selectedMBTilesFile.equals(mapActivity.getString(R.string.hotOSMOptionTitle))) {
+
+                        addOnlineDataSources();
+
+                    } else {
+
+                        addOfflineDataSources(selectedMBTilesFile);
+                    }
+                }
+            });
+
+            //handle cancel button tap event of dialog
+            builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int id) {
+                    //user clicked cancel
+                }
+            });
+
+            //present dialog to user
+            builder.show();
+
+        } else {
+
+            Toast prompt = Toast.makeText(context, "Please add .mbtiles file to " + mapActivity.getString(R.string.mbtilesAppPath), Toast.LENGTH_LONG);
+            prompt.show();
+        }
+    }
+
+    /**
+     * For instantiating a map (when the device is offline) and initializing the default mbtiles layer, extent, and zoom level
+     */
+    private void addOfflineDataSources(String fileName) {
+        String filePath = Environment.getExternalStorageDirectory() + "/" + ExternalStorage.APP_DIR + "/" + ExternalStorage.MBTILES_DIR + "/";
+        if(ExternalStorage.isReadable()) {
+            //fetch mbtiles from application folder (e.g. openmapkit/mbtiles)
+            File targetMBTiles = ExternalStorage.fetchFileFromExternalStorage(filePath + fileName);
+
+            if(!targetMBTiles.exists()) {
+                //inform user if no mbtiles was found
+                AlertDialog.Builder builder = new AlertDialog.Builder(mapActivity);
+                builder.setTitle("Device is Offline");
+                builder.setMessage("Please add mbtiles to " + filePath);
+                builder.setPositiveButton("Ok", new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int id) {
+                        //placeholder
+                    }
+                });
+
+                AlertDialog dialog = builder.create();
+                dialog.show();
+
+            } else {
+                //add mbtiles to map
+                mapView.setTileSource(new MBTilesLayer(targetMBTiles));
+            }
+        }
+    }
+}
